revert(vim.ui.input): revert to old ui.input
I got rid of this for a while and realized that
its indispensable for me. When I brought it back,
I brought back an old one because I couldn't
find the most recent one (because the commit
history for this repo is a mess, my bad). So, I
found and am bringing back my favorite version of
this, from 14ed0d5