fix(statuscol/showbreak): wraps don't end correctly when line is long
Problem:
When the last wrapped line is the exact length where the cursor starts
on the next row if it is at the end of the line, that line isn't
recognized as the last wrapped line in the group.

Solution:
Subtract one from the line length value so used in the wrap calculation.